Реализован загрузчикы: add_tariff, add_all. Добавлен эндпойнт: api/v1/services/popular/.
